Change latitude and longitude to double

+++ WeddingWebsite/Models/Location.cs
@@ -4,8 +4,8 @@
 /// Latitude and longitude for use in automated things like maps.
 /// </summary>
 public record Location(
-    long Latitude,
-    long Longitude
+    double Latitude,
+    double Longitude
 )
 {
     public string GetGoogleMapsLink() => $"https://www.google.com/maps/search/?api=1&query={Latitude},{Longitude}";
